Add fractional floating surface spacing

+++ src/main/scala/com/serenity/state/manager/StateManagerEventPipeline.scala
@@ -1285,7 +1285,7 @@ final private[manager] class StateManagerEventPipeline(
         menu.selectedIndex,
         hasHeader = true,
         hasFooter = menu.items.nonEmpty,
-        itemGapRows = state.config.commandRunnerItemGapRows
+        itemGapRows = state.config.commandRunnerItemGapRows.toInt
       )
     yield (surface, menu, index)
 
@@ -1548,7 +1548,7 @@ final private[manager] class StateManagerEventPipeline(
             runner.selectedIndex,
             hasHeader = true,
             hasFooter = runner.visibleItems.nonEmpty || runner.statusMessage.nonEmpty,
-            itemGapRows = state.config.commandRunnerItemGapRows
+            itemGapRows = state.config.commandRunnerItemGapRows.toInt
           )
             .map(RunnerSelectVisibleItem(_))
         case SurfaceContent.CommandPaletteSubmenu(runner, groupId, previewOnly) =>
@@ -1568,7 +1568,7 @@ final private[manager] class StateManagerEventPipeline(
             hasHeader = group.nonEmpty,
             hasFooter = items.nonEmpty || runner.statusMessage.nonEmpty,
             reservedContentRows = detailRows,
-            itemGapRows = state.config.commandRunnerItemGapRows
+            itemGapRows = state.config.commandRunnerItemGapRows.toInt
           ).map { index =>
             if previewOnly then RunnerSelectPreviewSubmenuItem(groupId, index)
             else RunnerSelectSubmenuItem(index)
